refactor(services): реорганизован слой сервисов по доменным папкам
Файлы сервисов перенесены в группы Articles, Feeds, Networking, Queries и Repositories.
Добавлен ArticleQueryService и обновлены точки композиции приложения под новую структуру, чтобы улучшить навигацию по проекту и сделать границы сервисного слоя более явными.
