+++ RSSReader/ContentView.swift
@@ -2,13 +2,7 @@ import SwiftUI
 
 struct ContentView: View {
     var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
+        AppComposition.makeRoot(models: AppComposition.appModels)
     }
 }
