+++ RSSReader/Services/Queries/ArticleListFilter.swift
