+++ RSSReader/Services/Queries/ArticleListItemDTO.swift
