+++ RSSReader/Services/Queries/ReaderArticleDTO.swift
